[Behat] Notifcation context extraction
