+++ etc/behat/suites/ui/addressing/managing_countries.yml
@@ -14,5 +14,6 @@ default:
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.ui.admin.managing_countries
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_countries && @ui"
