+++ etc/behat/suites/ui/addressing/managing_zones.yml
@@ -20,5 +20,6 @@ default:
                 - sylius.behat.context.setup.zone
 
                 - sylius.behat.context.ui.admin.managing_zones
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_zones && @ui"
