+++ etc/behat/suites/ui/channel/managing_channels.yml
@@ -20,5 +20,6 @@ default:
                 - sylius.behat.context.setup.zone
 
                 - sylius.behat.context.ui.admin.managing_channels
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_channels && @ui"
