+++ etc/behat/suites/ui/currency/managing_currencies.yml
@@ -15,5 +15,6 @@ default:
                 - sylius.behat.context.setup.currency
 
                 - sylius.behat.context.ui.admin.managing_currencies
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_currencies && @ui"
