+++ etc/behat/suites/ui/locale/managing_locales.yml
@@ -13,5 +13,6 @@ default:
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.ui.admin.managing_locales
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_locales && @ui"
