+++ etc/behat/suites/ui/payment/managing_payment_methods.yml
@@ -16,5 +16,6 @@ default:
                 - sylius.behat.context.transform.shared_storage
 
                 - sylius.behat.context.ui.admin.managing_payment_methods
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_payment_methods && @ui"
