+++ etc/behat/suites/ui/product/managing_product_attributes.yml
@@ -16,5 +16,6 @@ default:
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.ui.admin.managing_product_attributes
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_product_attributes && @ui"
