+++ etc/behat/suites/ui/product/managing_product_options.yml
@@ -17,5 +17,6 @@ default:
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.ui.admin.managing_product_options
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_product_options && @ui"
