+++ etc/behat/suites/ui/promotion/managing_promotions.yml
@@ -27,5 +27,6 @@ default:
                 - sylius.behat.context.setup.taxonomy
 
                 - sylius.behat.context.ui.admin.managing_promotions
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_promotions && @ui"
