+++ etc/behat/suites/ui/shipping/managing_shipping_methods.yml
@@ -26,6 +26,6 @@ default:
                 - sylius.behat.context.setup.zone
 
                 - sylius.behat.context.ui.admin.managing_shipping_methods
-
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_shipping_methods && @ui"
