+++ etc/behat/suites/ui/taxation/managing_tax_categories.yml
@@ -14,5 +14,6 @@ default:
                 - sylius.behat.context.setup.taxation
 
                 - sylius.behat.context.ui.admin.managing_tax_categories
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_tax_categories && @ui"
