+++ etc/behat/suites/ui/taxation/managing_tax_rates.yml
@@ -17,5 +17,6 @@ default:
                 - sylius.behat.context.setup.zone
 
                 - sylius.behat.context.ui.admin.managing_tax_rate
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_tax_rates && @ui"
