+++ etc/behat/suites/ui/taxonomy/managing_taxons.yml
@@ -15,5 +15,6 @@ default:
                 - sylius.behat.context.setup.taxonomy
 
                 - sylius.behat.context.ui.admin.managing_taxons
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_taxons && @ui"
