+++ etc/behat/suites/ui/user/managing_customers.yml
@@ -15,5 +15,6 @@ default:
                 - sylius.behat.context.setup.user
 
                 - sylius.behat.context.ui.admin.managing_customers
+                - sylius.behat.context.ui.admin.notification
             filters:
                 tags: "@managing_customers && @ui"
