+++ features/locale/managing_locales/adding_locale.feature
@@ -12,5 +12,5 @@ Feature: Adding a new locale
         Given I want to create a new locale
         When I choose Norwegian
         And I add it
-        Then I should be notified about successful creation
+        Then I should be notified that it has been successfully created
         And the store should be available in the Norwegian language
