+++ features/taxation/managing_tax_categories/editing_tax_category.feature
@@ -26,5 +26,5 @@ Feature: Editing tax category
         Given I want to modify a tax category "Alcohol"
         When I rename it to "Food & Alcohol"
         And I save my changes
-        Then I should be notified about successful edition
+        Then I should be notified that it has been successfully edited
         And this tax category name should be "Food & Alcohol"
